changing the unit of speed/throat/brake from "int" to "float"

+++ lib/Actors/movable.cpp
@@ -3,28 +3,28 @@
 
 Movable::Movable(QObject* parent, QGraphicsItem* newGraphicItem) : Actor(parent, newGraphicItem)
 {
-    maxSpeed =0; //decimeters per tick (1 km/h = 2.7778 cm/s), tick = 200ms (0.2s)
-    actualSpeed = 0; //decimeters per tick (200ms)
+    maxSpeed =0.0f; //decimeters per tick (1 km/h = 2.7778 cm/s), tick = 200ms (0.2s)
+    actualSpeed = 0.0f; //decimeters per tick (200ms)
 }
 
-void Movable::setActualSpeed(int newSpeed)
+void Movable::setActualSpeed(float newSpeed)
 {
     if (newSpeed <= maxSpeed) actualSpeed = newSpeed;
     else actualSpeed = maxSpeed;
-    if (actualSpeed < 0) actualSpeed = 0;
+    if (actualSpeed < 0.0f) actualSpeed = 0.0f;
 }
 
-int Movable::getMaxSpeed()
+float Movable::getMaxSpeed()
 {
     return maxSpeed;
 }
 
-void Movable::setMaxSpeed(int newMaxSpeed)
+void Movable::setMaxSpeed(float newMaxSpeed)
 {
     maxSpeed = newMaxSpeed;
 }
 
-const int Movable::getActualSpeed()
+const float Movable::getActualSpeed()
 {
     return actualSpeed;
 }
